requirements.txt
Flask==3.0.0
flask-session==0.5.0
PyMySQL==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1